requirements.txt
# requirements.txt
selenium==4.12.0
webdriver-manager==4.0.0
mysql-connector-python==8.1.0
argparse==1.4.0
python-dotenv
